requirements.txt
# Core dependencies
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyyaml>=6.0.1

# CSV handling
pandas>=2.0.0
chardet>=5.2.0

# Progress tracking
tqdm>=4.66.0
rich>=13.7.0

# Async support
asyncio>=3.4.3
asyncpg>=0.29.0

# Utilities
click>=8.1.7
colorama>=0.4.6

# Performance monitoring
psutil>=5.9.8
memory-profiler>=0.61.0

# Error handling and retries
tenacity>=8.2.3

# Data validation
cerberus>=1.3.5